Rename ImmutableArray to ArrayImitator and change its behavior
Reorder all methods by name ASC
