+++ src/AbstractArray.php
@@ -17,9 +17,7 @@ use IteratorAggregate;
 use Traversable;
 
 /**
- * Class AbstractArray
- *
- * Defines common methods and method signatures of the Mutable- and ImmutableArray.
+ * Defines common methods and method signatures.
  *
  * @author Victor Bocharsky <bocharsky.bw@gmail.com>
  */
@@ -45,6 +43,11 @@ abstract class AbstractArray implements
     const DEFAULT_SEPARATOR = ', ';
 
     /**
+     * @const array
+     */
+    const EMPTY_ARRAY = [];
+
+    /**
      * @var array
      */
     protected $elements = [];
@@ -59,49 +62,7 @@ abstract class AbstractArray implements
         $this->elements = $elements;
     }
 
-    /**
-     * Reindex the array numerically.
-     *
-     * @return $this An array with numerically-indexed elements
-     */
-    abstract public function reindex();
-
-    /**
-     * Exchanges all array keys with their associated values.
-     *
-     * @return static An array with flipped elements
-     */
-    abstract public function flip();
-
-    /**
-     * Reverse the order of the array values.
-     *
-     * @param bool $preserveKeys Whether array keys are preserved or no
-     *
-     * @return $this An array with the order of the elements reversed
-     */
-    abstract public function reverse($preserveKeys = false);
-
-    /**
-     * Pad array to the specified size with a given value.
-     *
-     * @param int $size Size of the result array
-     * @param mixed $value Empty value by default
-     *
-     * @return static An array padded to $size with $value
-     */
-    abstract public function pad($size, $value);
-
-    /**
-     * Extract a slice of the array.
-     *
-     * @param int $offset Slice begin index
-     * @param int|null $length Length of the slice
-     * @param bool $preserveKeys Whether array keys are preserved or no
-     *
-     * @return static A slice of the original array with length $length
-     */
-    abstract public function slice($offset, $length = null, $preserveKeys = false);
+    // The abstract public method list order by ASC
 
     /**
      * Split array into chunks.
@@ -114,23 +75,81 @@ abstract class AbstractArray implements
     abstract public function chunk($size, $preserveKeys = false);
 
     /**
-     * Removes duplicate values from the array.
+     * Clear array
      *
-     * @param int|null $sortFlags
-     *
-     * @return static An array with only unique elements
+     * @return static An empty array.
      */
-    abstract public function unique($sortFlags = null);
+    abstract public function clear();
 
     /**
-     * Merges this array with the provided one. Latter array is overwriting.
+     * Create an array using this array as values and the other array as keys.
      *
-     * @param array $array Array to merge with (overwrites)
-     * @param bool $recursively Whether array will be merged recursively or no
+     * @param array $array Key array
      *
-     * @return static An array with the keys/values from $array added
+     * @return static An array with keys from the other.
      */
-    abstract public function mergeWith(array $array, $recursively = false);
+    abstract public function combineTo(array $array);
+
+    /**
+     * Create an array using this array as keys and the other array as values.
+     *
+     * @param array $array Values array
+     *
+     * @return static An array with values from the other array
+     */
+    abstract public function combineWith(array $array);
+
+    /**
+     * {@inheritdoc}
+     *
+     * @return $this
+     *
+     * @link http://php.net/manual/en/function.usort.php
+     */
+    abstract public function customSort(callable $func);
+
+    /**
+     * {@inheritdoc}
+     *
+     * @return $this
+     *
+     * @link http://php.net/manual/en/function.uksort.php
+     */
+    abstract public function customSortKeys(callable $func);
+
+    /**
+     * Compute the array of values not present in the other array.
+     *
+     * @param array $array Array for diff
+     *
+     * @return static An array containing all the entries from this array that are not present in $array
+     */
+    abstract public function diffWith(array $array);
+
+    /**
+     * Filter the array for elements satisfying the predicate $func.
+     *
+     * @param callable $func
+     *
+     * @return static An array with only element satisfying $func
+     */
+    abstract public function filter(callable $func);
+
+    /**
+     * Exchanges all array keys with their associated values.
+     *
+     * @return static An array with flipped elements
+     */
+    abstract public function flip();
+
+    /**
+     * Apply the given function to the every element of the array, collecting the results.
+     *
+     * @param callable $func
+     *
+     * @return static An array with modified elements
+     */
+    abstract public function map(callable $func);
 
     /**
      * Merges array with the provided one. This array is overwriting.
@@ -143,15 +162,31 @@ abstract class AbstractArray implements
     abstract public function mergeTo(array $array, $recursively = false);
 
     /**
-     * Replace values in this array with values in the other array that have the
-     * same key.
+     * Merges this array with the provided one. Latter array is overwriting.
      *
-     * @param array $array Array of replacing values
-     * @param bool $recursively Whether array will be replaced recursively or no
+     * @param array $array Array to merge with (overwrites)
+     * @param bool $recursively Whether array will be merged recursively or no
      *
-     * @return static An array with the same keys but new values
+     * @return static An array with the keys/values from $array added
      */
-    abstract public function replaceWith(array $array, $recursively = false);
+    abstract public function mergeWith(array $array, $recursively = false);
+
+    /**
+     * Pad array to the specified size with a given value.
+     *
+     * @param int $size Size of the result array
+     * @param mixed $value Empty value by default
+     *
+     * @return static An array padded to $size with $value
+     */
+    abstract public function pad($size, $value);
+
+    /**
+     * Reindex the array numerically.
+     *
+     * @return $this An array with numerically-indexed elements
+     */
+    abstract public function reindex();
 
     /**
      * Replace the entire array with the other one except keys present in both.
@@ -165,31 +200,24 @@ abstract class AbstractArray implements
     abstract public function replaceIn(array $array, $recursively = false);
 
     /**
-     * Create an array using this array as keys and the other array as values.
+     * Replace values in this array with values in the other array that have the
+     * same key.
      *
-     * @param array $array Values array
+     * @param array $array Array of replacing values
+     * @param bool $recursively Whether array will be replaced recursively or no
      *
-     * @return static An array with values from the other array
+     * @return static An array with the same keys but new values
      */
-    abstract public function combineWith(array $array);
+    abstract public function replaceWith(array $array, $recursively = false);
 
     /**
-     * Create an array using this array as values and the other array as keys.
+     * Reverse the order of the array values.
      *
-     * @param array $array Key array
+     * @param bool $preserveKeys Whether array keys are preserved or no
      *
-     * @return static An array with keys from the other.
+     * @return $this An array with the order of the elements reversed
      */
-    abstract public function combineTo(array $array);
-
-    /**
-     * Compute the array of values not present in the other array.
-     *
-     * @param array $array Array for diff
-     *
-     * @return static An array containing all the entries from this array that are not present in $array
-     */
-    abstract public function diffWith(array $array);
+    abstract public function reverse($preserveKeys = false);
 
     /**
      * Randomize element order.
@@ -199,9 +227,20 @@ abstract class AbstractArray implements
     abstract public function shuffle();
 
     /**
+     * Extract a slice of the array.
+     *
+     * @param int $offset Slice begin index
+     * @param int|null $length Length of the slice
+     * @param bool $preserveKeys Whether array keys are preserved or no
+     *
+     * @return static A slice of the original array with length $length
+     */
+    abstract public function slice($offset, $length = null, $preserveKeys = false);
+
+    /**
      * {@inheritdoc}
      *
-     * @return MutableArray
+     * @return $this
      *
      * @link http://php.net/manual/en/function.arsort.php
      * @link http://php.net/manual/en/function.sort.php
@@ -213,7 +252,7 @@ abstract class AbstractArray implements
     /**
      * {@inheritdoc}
      *
-     * @return MutableArray
+     * @return $this
      *
      * @link http://php.net/manual/en/function.ksort.php
      * @link http://php.net/manual/en/function.krsort.php
@@ -221,22 +260,13 @@ abstract class AbstractArray implements
     abstract public function sortKeys($order = SORT_ASC, $strategy = SORT_REGULAR);
 
     /**
-     * Apply the given function to the every element of the array, collecting the results.
+     * Removes duplicate values from the array.
      *
-     * @param callable $func
+     * @param int|null $sortFlags
      *
-     * @return static An array with modified elements
+     * @return static An array with only unique elements
      */
-    abstract public function map(callable $func);
-
-    /**
-     * Filter the array for elements satisfying the predicate $func.
-     *
-     * @param callable $func
-     *
-     * @return static An array with only element satisfying $func
-     */
-    abstract public function filter(callable $func);
+    abstract public function unique($sortFlags = null);
 
     /**
      * Apply the given function to every element in the array, discarding the results.
@@ -248,30 +278,7 @@ abstract class AbstractArray implements
      */
     abstract public function walk(callable $func, $recursively = false);
 
-    /**
-     * {@inheritdoc}
-     *
-     * @return MutableArray
-     *
-     * @link http://php.net/manual/en/function.usort.php
-     */
-    abstract public function customSort(callable $func);
-
-    /**
-     * {@inheritdoc}
-     *
-     * @return MutableArray
-     *
-     * @link http://php.net/manual/en/function.uksort.php
-     */
-    abstract public function customSortKeys(callable $func);
-
-    /**
-     * Clear array
-     *
-     * @return static An empty array.
-     */
-    abstract public function clear();
+    // The public static method list order by ASC
 
     /**
      * Create a new instance.
@@ -283,6 +290,20 @@ abstract class AbstractArray implements
     public static function create(array $elements = [])
     {
         return new static($elements);
+    }
+
+    /**
+     * Decode a JSON string to new instance.
+     *
+     * @param string $json The JSON string being decoded
+     * @param int $options Bitmask of JSON decode options
+     * @param int $depth Specified recursion depth
+     *
+     * @return static The created array
+     */
+    public static function createFromJson($json, $options = 0, $depth = 512)
+    {
+        return new static(json_decode($json, true, $depth, $options));
     }
 
     /**
@@ -301,20 +322,6 @@ abstract class AbstractArray implements
         }
 
         return $array;
-    }
-
-    /**
-     * Decode a JSON string to new instance.
-     *
-     * @param string $json The JSON string being decoded
-     * @param int $options Bitmask of JSON decode options
-     * @param int $depth Specified recursion depth
-     *
-     * @return static The created array
-     */
-    public static function createFromJson($json, $options = 0, $depth = 512)
-    {
-        return new static(json_decode($json, true, $depth, $options));
     }
 
     /**
@@ -344,65 +351,7 @@ abstract class AbstractArray implements
         return new static(range($low, $high, $step));
     }
 
-    /**
-     * Clone current instance to new instance.
-     *
-     * @return $this Shallow copy of $this
-     */
-    public function createClone()
-    {
-        return clone $this;
-    }
-
-    /**
-     * Check whether the array is empty or not.
-     *
-     * @return bool Returns true if empty, false otherwise
-     */
-    public function isEmpty()
-    {
-        return !$this->elements;
-    }
-
-    /**
-     * Check whether array is associative or not.
-     *
-     * @return bool Returns true if associative, false otherwise
-     */
-    public function isAssoc()
-    {
-        if ($this->isEmpty()) {
-            return false;
-        }
-
-        foreach ($this->getKeys() as $key) {
-            if (!is_string($key)) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Check whether array is numeric or not.
-     *
-     * @return bool Returns true if numeric, false otherwise
-     */
-    public function isNumeric()
-    {
-        if ($this->isEmpty()) {
-            return false;
-        }
-
-        foreach ($this->getKeys() as $key) {
-            if (!is_int($key)) {
-                return false;
-            }
-        }
-
-        return true;
-    }
+    // The public method list order by ASC
 
     /**
      * Check if the given value exists in the array.
@@ -429,6 +378,30 @@ abstract class AbstractArray implements
     }
 
     /**
+     * Returns the number of values in the array.
+     *
+     * @link http://php.net/manual/en/function.count.php
+     *
+     * @return int total number of values
+     */
+    public function count()
+    {
+        return count($this->elements);
+    }
+
+    /**
+     * Clone current instance to new instance.
+     *
+     * @deprecated Should be removed
+     *
+     * @return $this Shallow copy of $this
+     */
+    public function createClone()
+    {
+        return clone $this;
+    }
+
+    /**
      * Find the given value in the array using a closure
      *
      * @param callable $func
@@ -447,15 +420,55 @@ abstract class AbstractArray implements
     }
 
     /**
-     * Search for a given element and return the index of its first occurrence.
+     * Returns the first occurrence of a value that satisfies the predicate $func.
      *
-     * @param mixed $element Value to search for
+     * @param callable $func
      *
-     * @return mixed The corresponding key/index
+     * @return mixed The first occurrence found
      */
-    public function indexOf($element)
+    public function find(callable $func)
     {
-        return array_search($element, $this->elements, true);
+        foreach ($this->elements as $key => $value) {
+            if($func($value, $key)) {
+                return $value;
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * Create an iterator over this array.
+     *
+     * @link http://php.net/manual/en/iteratoraggregate.getiterator.php
+     *
+     * @return Traversable An instance of an object implementing <b>Iterator</b>
+     */
+    public function getIterator()
+    {
+        return new ArrayIterator($this->elements);
+    }
+
+    /**
+     * Return an array all the keys of this array.
+     *
+     * @return array An array of all keys
+     */
+    public function getKeys()
+    {
+        return array_keys($this->elements);
+    }
+
+    /**
+     * Pick a random value out of this array.
+     *
+     * @return mixed Random value of array
+     *
+     * @throws \LogicException
+     */
+    public function getRandom()
+    {
+        return $this->offsetGet($this->getRandomKey());
     }
 
     /**
@@ -474,18 +487,6 @@ abstract class AbstractArray implements
         }
 
         return array_rand($this->elements, 1);
-    }
-
-    /**
-     * Pick a random value out of this array.
-     *
-     * @return mixed Random value of array
-     *
-     * @throws \LogicException
-     */
-    public function getRandom()
-    {
-        return $this->offsetGet($this->getRandomKey());
     }
 
     /**
@@ -534,21 +535,12 @@ abstract class AbstractArray implements
     public function getRandomValues($number)
     {
         $values = [];
+
         foreach ($this->getRandomKeys($number) as $key) {
             $values[] = $this->offsetGet($key);
         }
 
         return $values;
-    }
-
-    /**
-     * Return an array all the keys of this array.
-     *
-     * @return array An array of all keys
-     */
-    public function getKeys()
-    {
-        return array_keys($this->elements);
     }
 
     /**
@@ -562,28 +554,65 @@ abstract class AbstractArray implements
     }
 
     /**
-     * Reduce the array to a single value iteratively combining all values using $func.
+     * Search for a given element and return the index of its first occurrence.
      *
-     * @param callable $func callback ($carry, $item) -> next $carry
-     * @param mixed|null $initial starting value of the $carry
+     * @param mixed $element Value to search for
      *
-     * @return mixed Final value of $carry
+     * @return mixed The corresponding key/index
      */
-    public function reduce(callable $func, $initial = null)
+    public function indexOf($element)
     {
-        return array_reduce($this->elements, $func, $initial);
+        return array_search($element, $this->elements, true);
     }
 
     /**
-     * Returns the number of values in the array.
+     * Check whether array is associative or not.
      *
-     * @link http://php.net/manual/en/function.count.php
-     *
-     * @return int total number of values
+     * @return bool Returns true if associative, false otherwise
      */
-    public function count()
+    public function isAssoc()
     {
-        return count($this->elements);
+        if ($this->isEmpty()) {
+            return false;
+        }
+
+        foreach ($this->getKeys() as $key) {
+            if (!is_string($key)) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    /**
+     * Check whether the array is empty or not.
+     *
+     * @return bool Returns true if empty, false otherwise
+     */
+    public function isEmpty()
+    {
+        return !$this->elements;
+    }
+
+    /**
+     * Check whether array is numeric or not.
+     *
+     * @return bool Returns true if numeric, false otherwise
+     */
+    public function isNumeric()
+    {
+        if ($this->isEmpty()) {
+            return false;
+        }
+
+        foreach ($this->getKeys() as $key) {
+            if (!is_int($key)) {
+                return false;
+            }
+        }
+
+        return true;
     }
 
     /**
@@ -651,15 +680,16 @@ abstract class AbstractArray implements
     }
 
     /**
-     * Create an iterator over this array.
+     * Reduce the array to a single value iteratively combining all values using $func.
      *
-     * @link http://php.net/manual/en/iteratoraggregate.getiterator.php
+     * @param callable $func callback ($carry, $item) -> next $carry
+     * @param mixed|null $initial starting value of the $carry
      *
-     * @return Traversable An instance of an object implementing <b>Iterator</b>
+     * @return mixed Final value of $carry
      */
-    public function getIterator()
+    public function reduce(callable $func, $initial = null)
     {
-        return new ArrayIterator($this->elements);
+        return array_reduce($this->elements, $func, $initial);
     }
 
     /**
@@ -705,23 +735,5 @@ abstract class AbstractArray implements
             default:
                 ksort($elements, $strategy);
         }
-    }
-
-    /**
-     * Returns the first occurrence of a value that satisfies the predicate $func.
-     *
-     * @param callable $func
-     *
-     * @return mixed The first occurrence found
-     */
-    public function find(callable $func)
-    {
-        foreach ($this->elements as $key => $value) {
-            if($func($value, $key)) {
-                return $value;
-            }
-        }
-
-        return null;
     }
 }
